+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Fixed
+
+- Fixed changes to stdout not being printed after each I/O copy when using `pueue follow`.
+
 ## [3.1.2] - unreleased
 
 ## Fixed